updated help to separate sections out into individual components
